earnapp-api/app/requirements.txt
Flask
requests
pyyaml